Double precision

+++ class/InputHub.h
@@ -16,8 +16,8 @@ class IInputListener
 {
 public:
 	virtual void ProcessKey(KeyState keyState, uint8_t keyCode) = 0;
-	virtual void ProcessMouse(KeyState keyState, const Vector2f& mousePosition) = 0;
-	virtual void ProcessMouse(const Vector2f& mousePosition) = 0;
+	virtual void ProcessMouse(KeyState keyState, const Vector2d& mousePosition) = 0;
+	virtual void ProcessMouse(const Vector2d& mousePosition) = 0;
 };
 
 class InputHub : public Singleton<InputHub>
@@ -27,8 +27,8 @@ public:
 
 public:
 	void ProcessKey(KeyState keyState, uint8_t keyCode);
-	void ProcessMouse(KeyState keyState, const Vector2f& mousePosition);
-	void ProcessMouse(const Vector2f& mousePosition);
+	void ProcessMouse(KeyState keyState, const Vector2d& mousePosition);
+	void ProcessMouse(const Vector2d& mousePosition);
 
 public:
 	void Register(const std::shared_ptr<IInputListener>& pListener);
