+++ component/Character.h
@@ -26,8 +26,8 @@ public:
 
 	// For input listener
 	void ProcessKey(KeyState keyState, uint8_t keyCode) override;
-	void ProcessMouse(KeyState keyState, const Vector2f& mousePosition) override;
-	void ProcessMouse(const Vector2f& mousePosition) override;
+	void ProcessMouse(KeyState keyState, const Vector2d& mousePosition) override;
+	void ProcessMouse(const Vector2d& mousePosition) override;
 
 	void Move(uint32_t dir, float delta);
